state saving: use a tempfile::NamedTempFile to avoid file name conflicts
now when multiple processes/computers want to create the same shared
state file, they all write into different files and move them (in the
same) place afterwards. So one process "wins" and provides the shared
state for resuming, instead of corrupting the file by concurrent writes.
The problem was that even if the precomputed shared state for the same
graph is equivalent every time, the representation containing a
(randomized) hash map is not.